+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.5.13
 + Added a config option to enable TE Machine Frames in the recipes
++ Added a blacklist for the mob duplicator, blacklisted entities are shown when captured in the Mob Imprisonment Tools
 
 # 1.5.12
 + Fixed BHU being derpy with the new ItemHandler for big stacks